получение авторизации из контекста
