+++ src/components/App/model/ThemeContext.ts
@@ -1,0 +1,12 @@
+import {createContext} from "react";
+
+export interface ThemeContextType {
+    theme: string;
+    toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType>({
+    theme: 'light',
+    toggleTheme: () => {
+    }
+});
