+++ src/components/Dish/ui/Dish.tsx
@@ -1,33 +1,32 @@
 import {type FC, useContext} from "react";
-import { Counter } from "../../Counter";
-import { useCount } from "../../Counter/model/useCount";
-
+import {Counter} from "../../Counter";
+import {useCount} from "../../Counter/model/useCount";
+import {UserContext} from "../../App/model/UserContext.ts";
 import styles from "./Dish.module.css";
-import {UserContext} from "../../App/model/UserContext.tsx";
 
 
 interface DishProps {
-  dishName: string;
+    dishName: string;
 }
-export const Dish: FC<DishProps> = ({ dishName }) => {
-  const { count, onDecrement, onIncrement } = useCount();
-   const {user: {isAuthenticated} } = useContext(UserContext)
 
+export const Dish: FC<DishProps> = ({dishName}) => {
+    const {count, onDecrement, onIncrement} = useCount();
+    const {user: {isAuthenticated}} = useContext(UserContext)
 
-  return (
-    <>
-      <li className={styles.wrapper}>
-        <span className={styles.name}>{dishName}</span>
-        {isAuthenticated && <div className={styles.quantity}>
-          <Counter
-              count={count}
-              onDecrement={onDecrement}
-              onIncrement={onIncrement}
-          />
-        </div>}
-      </li>
-    </>
-  );
+    return (
+        <>
+            <li className={styles.wrapper}>
+                <span className={styles.name}>{dishName}</span>
+                {isAuthenticated && <div className={styles.quantity}>
+                    <Counter
+                        count={count}
+                        onDecrement={onDecrement}
+                        onIncrement={onIncrement}
+                    />
+                </div>}
+            </li>
+        </>
+    );
 };
