+++ src/components/Header/ui/Header.tsx
@@ -1,15 +1,13 @@
 import {forwardRef, useContext} from "react";
 import styles from "./Header.module.css";
 import {Button} from "../../Button";
-import {ThemeContext} from "../../App/model/ThemeContext.tsx";
-import {UserContext} from "../../App/model/UserContext.tsx";
+import {UserContext} from "../../App/model/UserContext.ts";
+import {ThemeContext} from "../../App/model/ThemeContext.ts";
 
 
 export const Header = forwardRef<HTMLDivElement>((_, ref) => {
-    const { toggleTheme} = useContext(ThemeContext)
+    const {toggleTheme} = useContext(ThemeContext)
     const {user, toggleUser} = useContext(UserContext)
-
-
 
 
     return (
@@ -17,7 +15,8 @@ export const Header = forwardRef<HTMLDivElement>((_, ref) => {
             <span>Header</span>
             <div className={styles.button_wrapper}>
                 <Button className={styles.button} title={'Theme'} onClick={toggleTheme}/>
-                {!user.isAuthenticated && <Button className={styles.button} title={'Вход'} onClick={() => toggleUser()}/>}
+                {!user.isAuthenticated &&
+                    <Button className={styles.button} title={'Вход'} onClick={() => toggleUser()}/>}
                 {user.userName && <Button className={styles.button} title={'Выход'} onClick={() => toggleUser()}/>}
             </div>
         </header>
